Set respawn standard streams using environment
The respawn applet gets and sets the standard streams from the following
environment variables: STDIN, STDOUT and STDERR. If unset, it uses the
null device.

It introduces the function __getenv(variable, undef) that returns the
variable from the environment if it is set; or the undef value if it is
not set.