Fix overlap while re-using strtoargv() in respawn
Commit 3a696bf (Re-use strtoargv() to serialize EXEC variable) has
introduced an overlap in respawn() due to the convertion of the exec
property in an argv array in pid_respawn().

The pid_respawn() converts the exec attribute in an argv array, which is
latter re-converted in string and finally reused in execve().

On the other hand, pid_respawn() must not modify the exec attribute.

Instead, it now makes a local copy which is then given to respawn().